Add mainstream category selection to org form
Same UI as adding the topics (e.g. a bad UI that just has a fixed number of dropdowns).  Note we've added :inverse_of declarations to the models to allow the validations to pass on create (when creating both the org and the relationship without inverse_of the org on the relationship doesn't exist according to rails).