Always save a slug, even for non-English editions
OK, so not having a slug set turns out to make things compicated on
front-end. Thinking on it, it actually makes more sense to save the
slug as the id for non-English editions, that way there is no
intrinsic difference between an English edition and a non-English one
as far as front-end is concerned.