Routing helpers handle non-English editions
We set the locale in the routing parameters for non-English editions
to ensure that the generated links always link to the correct
translation.

Note that this assumes non-English editions never have additional
translations, as setting the locale in this way forces it to what we
set, i.e. LocalisedUrlPathHelper will not override it with the current
locale in the public-facing controllers. If in future we decide that
non-English editions *can* have translations, we will have to rethink
how we generate these links so that they default to the edition's
primary locale, but still allow overrides for public-facing routing on
localised pages.