Tweak test stubbing to deal with LocalisedModel
Because we are loading editions as LocalisedModels, stubbing the can?
method with the actual instance no longer works. The simple fix is to
remove the parameter requirements on the stub as we're only interested
in verifying that limited access editions are not accessible, which in
this instance, falls outside the remit of the Authority stuff.